% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicApi.R
\name{.commonDetailEndpoint}
\alias{.commonDetailEndpoint}
\title{Handle common detail endpoints}
\usage{
.commonDetailEndpoint(type, id, metadataType = NULL,
  queryParameters = NULL)
}
\arguments{
\item{type}{Type of detail endpoint}

\item{id}{The internal CIViC ID}

\item{metadataType}{Optional type for base detail endpoints}

\item{metadataType}{Optional query parameters for base detail endpoints}
}
\description{
Handle common detail endpoints
}
