% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicApi.R
\name{.commonIndexEndpoint}
\alias{.commonIndexEndpoint}
\title{Handle common index endpoints}
\usage{
.commonIndexEndpoint(type, page, count)
}
\arguments{
\item{type}{Type of index endpoint}

\item{page}{the page number to retrieve}

\item{count}{the number of assertions to retrieve}
}
\description{
Handle common index endpoints
}
