% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicApi.R
\name{.verifyJsonResponse}
\alias{.verifyJsonResponse}
\title{Verify that the httr response is of type "application/json"}
\usage{
.verifyJsonResponse(response)
}
\arguments{
\item{response}{httr response}
}
\description{
Verify that the httr response is of type "application/json"
}
