% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicr.R
\name{getAllEvidenceItems}
\alias{getAllEvidenceItems}
\title{Get a list of evidence items}
\usage{
getAllEvidenceItems(page = 1, count = 25)
}
\arguments{
\item{page}{the page number to retrieve}

\item{count}{the number of evidence items to retrieve}
}
\value{
An S3 Object of type civic_api containing the content, url, and response
}
\description{
Retrieve all evidence items from the CIViC DB
}
\examples{
getAllEvidenceItems(count = 10)
getAllEvidenceItems(page = 2, count = 10)
}
\keyword{evidence}
\keyword{items}
