% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicr.R
\name{getVariantGroup}
\alias{getVariantGroup}
\title{Get a specific variant group}
\usage{
getVariantGroup(id)
}
\arguments{
\item{id}{Internal CIViC ID of the variant group of interest}
}
\value{
An S3 Object of type civic_api containing the content, url, and response
}
\description{
Retrieve a specific variant group from the CIViC DB
}
\examples{
getVariantGroup(id = 1)
}
\keyword{group}
\keyword{variant}
