% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicApi.R
\name{getVariantGroupMetadata}
\alias{getVariantGroupMetadata}
\title{Get a specific variant group metadata information}
\usage{
getVariantGroupMetadata(id, type)
}
\arguments{
\item{id}{Internal CIViC ID of the variant group of interest}

\item{type}{Type of metadata (comments, suggested_changes, revisions)}
}
\value{
An S3 Object of type civic_api containing the content, url, and response
}
\description{
Retrieve metadata information for a specific variant group from the CIViC DB
}
\examples{
getVariantGroupMetadata(id = 1, type = "comments")
getVariantGroupMetadata(id = 1, type = "revisions")
}
\keyword{group,}
\keyword{metadata}
\keyword{variant}
